[SPARK-52871][CORE] Merge `o.a.s.sql.catalyst.util.SparkStringUtils` to `o.a.s.util.SparkStringUtils`
