+++ sql/api/src/main/scala/org/apache/spark/sql/catalyst/util/StringUtils.scala
@@ -16,12 +16,7 @@
  */
 package org.apache.spark.sql.catalyst.util
 
-import java.util.HexFormat
-import java.util.concurrent.atomic.AtomicBoolean
-
-import org.apache.spark.internal.Logging
 import org.apache.spark.unsafe.array.ByteArrayUtils
-import org.apache.spark.util.ArrayImplicits._
 
 /**
  * Concatenation of sequence of strings to final string with cheap append method and one memory
@@ -66,76 +61,3 @@ class StringConcat(val maxLength: Int = ByteArrayUtils.MAX_ROUNDED_ARRAY_LENGTH)
   }
 }
 
-object SparkStringUtils extends Logging {
-
-  /** Whether we have warned about plan string truncation yet. */
-  private val truncationWarningPrinted = new AtomicBoolean(false)
-
-  /**
-   * Format a sequence with semantics similar to calling .mkString(). Any elements beyond
-   * `maxFields` will be dropped and replaced by a "... N more fields" placeholder.
-   *
-   * @return
-   *   the trimmed and formatted string.
-   */
-  def truncatedString[T](
-      seq: Seq[T],
-      start: String,
-      sep: String,
-      end: String,
-      maxFields: Int,
-      customToString: Option[T => String] = None): String = {
-    if (seq.length > maxFields) {
-      if (truncationWarningPrinted.compareAndSet(false, true)) {
-        logWarning(
-          "Truncated the string representation of a plan since it was too large. This " +
-            s"behavior can be adjusted by setting 'spark.sql.debug.maxToStringFields'.")
-      }
-      val numFields = math.max(0, maxFields)
-      val restNum = seq.length - numFields
-      val ending = (if (numFields == 0) "" else sep) +
-        (if (restNum == 0) "" else s"... $restNum more fields") + end
-      if (customToString.isDefined) {
-        seq.take(numFields).map(customToString.get).mkString(start, sep, ending)
-      } else {
-        seq.take(numFields).mkString(start, sep, ending)
-      }
-    } else {
-      if (customToString.isDefined) {
-        seq.map(customToString.get).mkString(start, sep, end)
-      } else {
-        seq.mkString(start, sep, end)
-      }
-    }
-  }
-
-  /** Shorthand for calling truncatedString() without start or end strings. */
-  def truncatedString[T](seq: Seq[T], sep: String, maxFields: Int): String = {
-    truncatedString(seq, "", sep, "", maxFields)
-  }
-
-  private final lazy val SPACE_DELIMITED_UPPERCASE_HEX =
-    HexFormat.of().withDelimiter(" ").withUpperCase()
-
-  /**
-   * Returns a pretty string of the byte array which prints each byte as a hex digit and add
-   * spaces between them. For example, [1A C0].
-   */
-  def getHexString(bytes: Array[Byte]): String = {
-    s"[${SPACE_DELIMITED_UPPERCASE_HEX.formatHex(bytes)}]"
-  }
-
-  def sideBySide(left: String, right: String): Seq[String] = {
-    sideBySide(left.split("\n").toImmutableArraySeq, right.split("\n").toImmutableArraySeq)
-  }
-
-  def sideBySide(left: Seq[String], right: Seq[String]): Seq[String] = {
-    val maxLeftSize = left.map(_.length).max
-    val leftPadded = left ++ Seq.fill(math.max(right.size - left.size, 0))("")
-    val rightPadded = right ++ Seq.fill(math.max(left.size - right.size, 0))("")
-
-    leftPadded.zip(rightPadded).map { case (l, r) =>
-      (if (l == r) " " else "!") + l + (" " * ((maxLeftSize - l.length) + 3)) + r
-    }
-  }
-}
