+++ sql/api/src/main/scala/org/apache/spark/sql/types/StructType.scala
@@ -30,11 +30,11 @@ import org.apache.spark.annotation.Stable
 import org.apache.spark.sql.catalyst.analysis.SqlApiAnalysis
 import org.apache.spark.sql.catalyst.parser.{DataTypeParser, LegacyTypeStringParser}
 import org.apache.spark.sql.catalyst.trees.Origin
-import org.apache.spark.sql.catalyst.util.{CaseInsensitiveMap, SparkStringUtils, StringConcat}
+import org.apache.spark.sql.catalyst.util.{CaseInsensitiveMap, StringConcat}
 import org.apache.spark.sql.errors.DataTypeErrors
 import org.apache.spark.sql.errors.DataTypeErrors.toSQLId
 import org.apache.spark.sql.internal.SqlApiConf
-import org.apache.spark.util.SparkCollectionUtils
+import org.apache.spark.util.{SparkCollectionUtils, SparkStringUtils}
 
 /**
  * A [[StructType]] object can be constructed by
